`refactor: restructure packages under 'payment' namespace`
Detailed body (optional):
- Moved all classes and interfaces to the `payment` namespace for consistency and better organization.
- Updated package declarations and imports across the project.
- Adjusted annotations in `ApplicationFactoryMethod` to reflect the new package structure.
- Removed unused or redundant code (e.g., `GroupedOpenApi` bean in `OpenApiConfig`).
- Improved JavaDoc comments for readability and documentation clarity.
