+++ src/main/java/payment/config/OpenApiConfig.java
@@ -1,11 +1,10 @@
-package config;
+package payment.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.info.Contact;
+import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.info.License;
 import io.swagger.v3.oas.models.servers.Server;
-import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -33,13 +32,4 @@ public class OpenApiConfig {
                                 .description("Development Server")
                 ));
     }
-
-    @Bean
-    public GroupedOpenApi paymentApi() {
-        return GroupedOpenApi.builder()
-                .group("payment-api")
-                .pathsToMatch("/api/payments/**")
-                .build();
-    }
-
 }
