+++ src/main/java/payment/exception/UnsupportedPaymentMethodException.java
@@ -1,4 +1,4 @@
-package exception;
+package payment.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
