+++ src/main/java/payment/model/PaymentStatus.java
@@ -1,4 +1,4 @@
-package model;
+package payment.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
