+++ src/main/java/payment/model/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
-package model.repository;
+package payment.model.repository;
 
-import model.entities.PaymentEntity;
+import payment.model.entities.PaymentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
