+++ src/main/java/payment/request/CreditCardPaymentRequest.java
@@ -1,4 +1,4 @@
-package request;
+package payment.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
