+++ src/main/java/payment/service/PaymentProcessFactory.java
@@ -1,13 +1,13 @@
-package service;
+package payment.service;
 
 import org.springframework.stereotype.Service;
-import request.BankSlipPaymentRequest;
-import request.CreditCardPaymentRequest;
-import request.PaymentRequest;
-import request.PixPaymentRequest;
-import service.impl.BankSlipPaymentProcessorImpl;
-import service.impl.CreditCardPaymentProcessorImpl;
-import service.impl.PixPaymentProcessor;
+import payment.request.BankSlipPaymentRequest;
+import payment.request.CreditCardPaymentRequest;
+import payment.request.PaymentRequest;
+import payment.request.PixPaymentRequest;
+import payment.service.impl.BankSlipPaymentProcessorImpl;
+import payment.service.impl.CreditCardPaymentProcessorImpl;
+import payment.service.impl.PixPaymentProcessor;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -20,13 +20,12 @@ public class PaymentProcessFactory {
     public PaymentProcessFactory(CreditCardPaymentProcessorImpl creditCardPaymentProcessorImpl,
                                  BankSlipPaymentProcessorImpl bankSlipPaymentProcessorImpl,
                                  PixPaymentProcessor pixPaymentProcessor) {
-        // Registra os processadores no mapa
         processors.put(CreditCardPaymentRequest.class, creditCardPaymentProcessorImpl);
         processors.put(BankSlipPaymentRequest.class, bankSlipPaymentProcessorImpl);
         processors.put(PixPaymentRequest.class, pixPaymentProcessor);
     }
 
-    // Factory Method
+    // FÁBRICA DE MÉTODOS DE PROCESSAMENTO DE PAGAMENTO
     public PaymentProcessorService createProcessor(PaymentRequest paymentRequest) {
         PaymentProcessorService processor = processors.get(paymentRequest.getClass());
