+++ src/main/java/payment/service/impl/CreditCardPaymentProcessorImpl.java
@@ -1,15 +1,15 @@
-package service.impl;
+package payment.service.impl;
 
-import model.PaymentMethod;
-import model.PaymentStatus;
-import model.entities.PaymentEntity;
-import model.repository.PaymentRepository;
+import payment.model.PaymentMethod;
+import payment.model.PaymentStatus;
+import payment.model.entities.PaymentEntity;
+import payment.model.repository.PaymentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import request.CreditCardPaymentRequest;
-import request.PaymentRequest;
-import response.PaymentResponse;
-import service.PaymentProcessorService;
+import payment.request.CreditCardPaymentRequest;
+import payment.request.PaymentRequest;
+import payment.response.PaymentResponse;
+import payment.service.PaymentProcessorService;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
